+++ jael/CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.0.2
+* Use `package:dart2_constant`.
+* Upgrade `package:symbol_table`.
+* Added `Renderer.errorDocument`.
+
 # 1.0.1
 * Reworked the scanner; thereby fixing an extremely pesky bug
 that prevented successful parsing of Jael files containing